integrate limit request into Support
