+++ src/screens/support-issue.screen.tsx
@@ -1,4 +1,14 @@
-import { ApiError, KycLevel, Utils, Validations, useSupport } from '@dfx.swiss/react';
+import {
+  ApiError,
+  FundOrigin,
+  InvestmentDate,
+  KycLevel,
+  Limit,
+  Utils,
+  Validations,
+  useSupport,
+  useUserContext,
+} from '@dfx.swiss/react';
 import {
   Form,
   StyledButton,
@@ -14,7 +24,14 @@ import { useForm, useWatch } from 'react-hook-form';
 import { useLocation, useParams } from 'react-router-dom';
 import { ErrorHint } from '../components/error-hint';
 import { Layout } from '../components/layout';
-import { IssueReasonLabels, IssueTypeLabels } from '../config/labels';
+import {
+  DateLabels,
+  IssueReasonLabels,
+  IssueTypeLabels,
+  LimitLabels,
+  OriginFutureLabels,
+  OriginNowLabels,
+} from '../config/labels';
 import { useSettingsContext } from '../contexts/settings.context';
 import { useKycLevelGuard, useUserGuard } from '../hooks/guard.hook';
 import { useNavigation } from '../hooks/navigation.hook';
@@ -34,6 +51,10 @@ interface FormData {
   transaction: string;
   reason: SupportIssueReason;
   message: string;
+  limit: Limit;
+  investmentDate: InvestmentDate;
+  fundOrigin: FundOrigin;
+  fundOriginText?: string;
   file?: File;
 }
 
@@ -47,10 +68,16 @@ export default function SupportIssueScreen(): JSX.Element {
   const rootRef = useRef<HTMLDivElement>(null);
   const { createIssue } = useSupport();
   const { translate, translateError } = useSettingsContext();
+  const { user } = useUserContext();
+  const { search } = useLocation();
 
   const [isLoading, setIsLoading] = useState(false);
   const [error, setError] = useState<string>();
   const [issueCreated, setIssueCreated] = useState(false);
+
+  const params = new URLSearchParams(search);
+  const paramKycCode = params.get('code');
+  const kycCode = paramKycCode ?? user?.kyc.hash;
 
   const {
     control,
@@ -64,8 +91,9 @@ export default function SupportIssueScreen(): JSX.Element {
     },
   });
   const selectedType = useWatch({ control, name: 'type' });
+  const investmentDate = useWatch({ control, name: 'investmentDate' });
 
-  const types = Object.values(SupportIssueType).filter((t) => t !== SupportIssueType.LIMIT_REQUEST);
+  const types = Object.values(SupportIssueType);
   const reasons = IssueReasons[selectedType] ?? [];
 
   useEffect(() => {
@@ -79,8 +107,8 @@ export default function SupportIssueScreen(): JSX.Element {
   useEffect(() => {
     if (selectedType === SupportIssueType.TRANSACTION_ISSUE) {
       if (!id) navigate('/support/issue/tx');
-    } else {
-      navigate('/support/issue');
+    } else if (selectedType === SupportIssueType.LIMIT_REQUEST) {
+      if (!kycCode) navigate('/kyc');
     }
   }, [selectedType]);
 
@@ -99,6 +127,16 @@ export default function SupportIssueScreen(): JSX.Element {
 
       if (data.type === SupportIssueType.TRANSACTION_ISSUE && id) {
         request.transaction = { id: +id };
+      }
+
+      if (data.type === SupportIssueType.LIMIT_REQUEST && data.limit) {
+        (request as any).limitRequest = {
+          // TODO: Add limitRequest to packages
+          limit: data.limit,
+          investmentDate: data.investmentDate,
+          fundOrigin: data.fundOrigin,
+          fundOriginText: data.fundOriginText,
+        };
       }
 
       await createIssue(request);
@@ -120,6 +158,9 @@ export default function SupportIssueScreen(): JSX.Element {
     name: Validations.Required,
     reason: Validations.Required,
     message: Validations.Required,
+    limit: Validations.Required,
+    investmentDate: Validations.Required,
+    fundOrigin: Validations.Required,
   });
 
   return (
@@ -187,6 +228,55 @@ export default function SupportIssueScreen(): JSX.Element {
               placeholder={`${translate('screens/kyc', 'John')} ${translate('screens/kyc', 'Doe')}`}
               full
             />
+
+            {selectedType === SupportIssueType.LIMIT_REQUEST && (
+              <>
+                <StyledDropdown<Limit>
+                  rootRef={rootRef}
+                  label={translate('screens/limit', 'Investment volume')}
+                  items={Object.values(Limit).filter((i) => typeof i !== 'string') as number[]}
+                  labelFunc={(item) => LimitLabels[item]}
+                  name="limit"
+                  placeholder={translate('general/actions', 'Select...')}
+                  full
+                />
+
+                <StyledDropdown<InvestmentDate>
+                  rootRef={rootRef}
+                  label={translate('screens/limit', 'Investment date')}
+                  items={Object.values(InvestmentDate)}
+                  labelFunc={(item) => translate('screens/limit', DateLabels[item])}
+                  name="investmentDate"
+                  placeholder={translate('general/actions', 'Select...')}
+                  full
+                />
+
+                <StyledDropdown<FundOrigin>
+                  rootRef={rootRef}
+                  label={translate('screens/limit', 'Origin of funds')}
+                  items={Object.values(FundOrigin)}
+                  labelFunc={(item) =>
+                    translate(
+                      'screens/limit',
+                      investmentDate === InvestmentDate.FUTURE ? OriginFutureLabels[item] : OriginNowLabels[item],
+                    )
+                  }
+                  name="fundOrigin"
+                  placeholder={translate('general/actions', 'Select...')}
+                  full
+                />
+
+                <StyledInput
+                  name="fundOriginText"
+                  label={`${translate('screens/limit', 'Origin of funds')} (${translate(
+                    'screens/limit',
+                    'free text',
+                  )})`}
+                  multiLine
+                  full
+                />
+              </>
+            )}
 
             <StyledInput name="message" label={translate('screens/support', 'Description')} multiLine full />
